reorganize project into modules
