+++ src/lib.rs
@@ -1,40 +1,11 @@
-// input parsers
-pub mod tsv_to_json;
-
-pub mod txt_to_json;
-
-pub mod parse_json;
-
-// processors
-pub mod add_fields;
-
-pub mod drop_fields;
-
-pub mod extract_items;
-
-pub mod join;
-
-pub mod merge;
-
-pub mod rename_fields;
-
-pub mod select_fields;
-
-pub mod to_number;
-
-// output converters
-pub mod output_json;
-
-pub mod output_tsv;
+pub mod pipeline;
+pub mod processor;
 
 // global stuff
-pub mod types;
-
 pub mod builders;
-
-pub mod util;
-
 mod json_lines_formatter;
+pub mod types;
+pub mod util;
 
 // test
 mod test;
