+++ src/processor/drop_fields.rs
