+++ src/processor/mod.rs
@@ -1,0 +1,15 @@
+pub mod add_fields;
+
+pub mod drop_fields;
+
+pub mod extract_items;
+
+pub mod join;
+
+pub mod merge;
+
+pub mod rename_fields;
+
+pub mod select_fields;
+
+pub mod to_number;
